+++ node/src/engine/shardchain.rs
@@ -1,4 +1,4 @@
-use crate::block::builder::{PreparedBlock, EngineTraceInfoData};
+use crate::block::builder::{EngineTraceInfoData, PreparedBlock};
 use crate::block::{BlockBuilder, BlockFinality};
 use crate::data::{DocumentsDb, NodeStorage, ShardStorage, ExternalAccountsProvider};
 use crate::engine::{BlockTimeMode, InMessagesQueue};
@@ -52,6 +52,10 @@ impl Shardchain {
         })
     }
 
+    pub(crate) fn nex_seq_no(&self) -> NodeResult<u32> {
+        Ok(self.block_finality.lock().get_last_info()?.0.seq_no() + 1)
+    }
+
     pub(crate) fn out_message_queue_is_empty(&self) -> bool {
         self.block_finality.lock().out_message_queue_is_empty()
     }
@@ -72,10 +76,7 @@ impl Shardchain {
             self.block_gas_limit,
             self.accounts_provider.clone(),
         )?;
-        collator.build_block(
-            &self.message_queue,
-            &self.blockchain_config,
-        )
+        collator.build_block(&self.message_queue, &self.blockchain_config)
     }
 
     ///
@@ -83,6 +84,7 @@ impl Shardchain {
     ///
     pub fn generate_block(
         &self,
+        mc_seq_no: u32,
         time: u32,
         time_mode: BlockTimeMode,
     ) -> NodeResult<Option<Block>> {
@@ -90,7 +92,12 @@ impl Shardchain {
         Ok(if !block.is_empty {
             log::trace!(target: "node", "block generated successfully");
             Self::print_block_info(&block.block);
-            self.finality_and_apply_block(block.block.clone(), block.state, block.transaction_traces)?;
+            self.finality_and_apply_block(
+                mc_seq_no,
+                block.block.clone(),
+                block.state,
+                block.transaction_traces,
+            )?;
             Some(block.block)
         } else {
             log::trace!(target: "node", "empty block was not generated");
@@ -111,15 +118,19 @@ impl Shardchain {
     /// finality and apply block
     pub(crate) fn finality_and_apply_block(
         &self,
+        mc_seq_no: u32,
         block: Block,
         applied_shard: ShardStateUnsplit,
         transaction_traces: HashMap<UInt256, Vec<EngineTraceInfoData>>,
     ) -> NodeResult<Arc<ShardStateUnsplit>> {
         log::info!(target: "node", "Apply block seq_no = {}", block.read_info()?.seq_no());
         let new_state = Arc::new(applied_shard);
-        self.block_finality
-            .lock()
-            .put_block_with_info(block, new_state.clone(), transaction_traces)?;
+        self.block_finality.lock().put_block_with_info(
+            mc_seq_no,
+            block,
+            new_state.clone(),
+            transaction_traces,
+        )?;
         Ok(new_state)
     }